Extracted meta inputs into separate component
